moving the contextio info into the helper module
