+++ app/assets/javascripts/angular/templates/pipeline_details.html
@@ -1,4 +1,4 @@
-<div class="row panel">
+<div class="row ">
   <div id="pipeDataDiv" class="large-8 columns">
     <!-- Displays title of the pipeline -->
     <div id="showDataDiv" ng-hide="editData">
@@ -32,6 +32,7 @@
 
     <h3>{{stage.name}}</h3>
     <p>{{stage.description}}</p>
+
     <table>
       <!-- Heading for fields in your pipeline -->
       <thead>
