+++ app/services/ContextIO/contextio_api_script.rb
@@ -1,3 +1,6 @@
+
+
+
 class ContextioApiScript < ApplicationController
 
 #this TX script runs all othe other contextio TX sscripts and passes on the errors
@@ -8,8 +11,14 @@ class ContextioApiScript < ApplicationController
     return contact unless contact.error.nil?
 
 #this authenticates the notification from contextio
-    contact = ApiAuthenticationScript.run(contact)
-    return contact unless contact.error.nil?
+    if inputs['signature'].nil?
+      return failure 'Context.io authentication signature is nil'
+    end
+
+  #the webhook id is obtained when we create the webhook
+    error = authenticate(inputs)
+
+    return failure error if !error.nil?
 
 #this retrieves the message body from contextio and returns the string containing the body
     contact = RetrieveMessageScript.run(contact)
